Job done

+++ src/com/pachetepachete/Models/Constraint.java
@@ -3,27 +3,27 @@ package com.pachetepachete.Models;
 import java.util.Date;
 
 public class Constraint {
-    private int start;
-    private int end;
+    private double start;
+    private double end;
 
-    public Constraint(int start, int end) {
+    public Constraint(double start, double end) {
         this.start = start;
         this.end = end;
     }
 
-    public int getStart() {
+    public double getStart() {
         return start;
     }
 
-    public void setStart(int start) {
+    public void setStart(double start) {
         this.start = start;
     }
 
-    public int getEnd() {
+    public double getEnd() {
         return end;
     }
 
-    public void setEnd(int end) {
+    public void setEnd(double end) {
         this.end = end;
     }
 }
